Cleaned and finished everything. Comments needed.

+++ C-System.ino
@@ -100,9 +100,9 @@ String FlightTimeStr() {
 
 
 
-long FlightTimeMinutes(){
+float FlightTimeMinutes(){
 
-  unsigned long minutes = millis() / 1000;
+  float minutes = millis() / 1000;
   minutes = minutes / 60;
   return minutes;
 }
